A quick hack to add config file support using argparse
gcovr will now search for a gcovr.conf file in --root (if given) and
current directory (in that order). If such a file is found it is read
and all arguments from the file is added before command line
arguments. So command line arguments have priority.

You can also use --config option to specify a config file.

The format of the config file is as follows:
# This is a comment
root = .
exclude = ../subprojects
exclude = ../test
print-summary

this is handled as if the command line arguments started with:
--root . --exclude ../subprojects --exclude ../test --print-summary